+++ packages/blockchain-ui/src/components/index.ts
@@ -1,6 +1,7 @@
 export { Button } from "./button";
 export * from "./flex";
 export * from "./charts";
+export * from "./nftCard";
 export * from "./input/input";
 export { Alert } from "./alert/alert";
 export * from "./typography/typography";
